server/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
